Don't specify # of batches, improve bscan glob & dataset batching
We do this because the images stored in xzInstensity are not all bscans
(eg. some of them are the enfaces generated using the bscans).
Instead of specifying the number of batches in a dataset, simply run
until the dataset is exhausted.
Instead of trying to manually remove batches that are less than the
desired batch size, simply enable that behaviour in the batch() call.