Add Final CNN files (#120)
* Update CNN

* Fix bugs

Some related to running in sharcnet, others are not.
Also added some short documenation.

* Adding enface generation and logging to Comet

* Remove Comet Logging

Will be replaced by Tensorboard logging.

* Remove explicit split between training datasets and testing datasets, restructure train/test methods

* Simplify enface generation

* Refactor

* Add logging and checkpoint saving back

* Restructure training/testing dataset generation

* Add enface image logging

* Refactor parameters file

* Log cross sections and enfaces after every epoch

* Increase shuffle buffer size

Its size must be at least equal to the number of samples
for perfect shuffling.

* Clean up import statements

* Update hyperparameters

To hopefully be more in line with the paper.
Also, minor clean up on the callback.

* Fix bugs related to running in sharcnet

* Don't specify # of batches, improve bscan glob & dataset batching

We do this because the images stored in xzInstensity are not all bscans
(eg. some of them are the enfaces generated using the bscans).
Instead of specifying the number of batches in a dataset, simply run
until the dataset is exhausted.
Instead of trying to manually remove batches that are less than the
desired batch size, simply enable that behaviour in the batch() call.

* Create new logs folder for each cnn.py execution

Before, attempting to continue training a model from a checkpoint
would result in the tensorboard graph jumping from some non-zero
step back to step zero.
Now, instead of adding onto the existing logs, a new logs folder
is created.

* Rename tensorboard enface images, increase shuffle buffer

Add the dataset name to the enface image name when logging to
Tensorboard.
Increased shuffle buffer to be at least 66 * 512 * NUM_SLICES big
(since there are currently 66 datasets with 512 images each).

* Change loss to MSE, simplify checkpoints

Removed stuff from the checkpoints that we don't actually need saved.

* Add code provided by Dr. Lee

* Update unet code to use tensorflow.keras

* Use u-net instead of building the model ourselves

This means changing the format of images from [H,W,C] to [C,H,W].

* Specify number of batches when training/testing/predicting

This is done to stop an 'end-of-sequence' error from appearing
after a dataset is exhausted.

Because we need to know exactly how many batches are in a dataset,
ignore_errors has been disabled.

* Fix num batches calculation

* Centre input images

Also removes the dataset blacklist, since those datasets are no
longer included in the root data directory.

Disabled enface mode because the training set mean and s.t.d. are
now needed to do predictions. Since the training set changes every
execution, the mean/s.t.d. are not consistent.

* Add standard output logging function

Removed some print statements that weren't needed, added
some others.
Also simplifies the naming of the tensorboard log dir.

* Fix issue where the data was split differently each time for the same seed

Had to do with the order of the datasets changing each time when we glob
for them.
Can re-enable enface mode now.

* Add print statements

To see how long calculating the mean and s.t.d. take.

* Fix enface generation in callback

Forgot to update callback function after changing how training/testing
split is done.

* Generate enfaces only every 5 epochs

* Allow num slices to be passed as command line arg

* Add separate script to calculate mean and std

* Add single_poly_flattened stats

* Fetch mean and std from parameters instead of calculating every time

* Add extra logging, fix issue with continually opening new plt figures

* Retain entire data dir paths, instead of only their names

* Rename variables

* Fix issue with black spots in saved images

* Allow images to be contrasted

Like in the cgan

* Convert image to float32

* If exception occurs during enface generation, skip eye

* Update stats script to work with k folds

Using a modified Datasets class, originally from the cgan.
Needed to be modified because the random sampling was not
consistent when using the same seed.

* Add k-folds training

Also to fixed a small bug in cnn.Dataset.

* Generate enfaces only at the end of training

Instead of once every 5 epochs

* Add k-folds mean and std

* Separate data loading into its own method

* Create script for generating initial cnn weights

To be used for creating the initial weights for k-folds training.

* Add script for aggregating scores from each fold

* Add normalize flag

* Add augmented data option

* Add different levels of data augmentation

* Pass angle value to image.load

Forgot to add this to an earlier commit

* Update documentation

Also removes some scripts that are out of date

* Add sharcnet changes